Flowers' Yasuo updated to latest
