+++ Core/Champion Ports/Yasuo/Flowers' Yasuo/Evade/Evader/EvadeTargetManager.cs
@@ -14,6 +14,7 @@ namespace Flowers_Yasuo.Evade
     internal class EvadeTargetManager // Credit By Brian
     {
         public static Menu Menu;
+        public static readonly List<SpellData> Spells = new List<SpellData>();
         private static Vector2 wallCastedPos;
         private static readonly List<Targets> DetectedTargets = new List<Targets>();
 
@@ -28,14 +29,204 @@ namespace Flowers_Yasuo.Evade
             }
         }
 
-        public static void Init()
+        public static void Init(Menu mainMenu)
         {
-            Menu = Logic.Menu;
+            Menu = mainMenu;
+
+            InitSpells();
+
+            var attack = Menu.AddSubMenu(new Menu("Dodge Attack", "Dodge Attack"));
+            {
+                attack.AddItem(new MenuItem("BAttack", "Basic Attack", true).SetValue(true));
+                attack.AddItem(new MenuItem("BAttackHpU", "-> If Hp <", true).SetValue(new Slider(35)));
+                attack.AddItem(new MenuItem("CAttack", "Crit Attack", true).SetValue(true));
+                attack.AddItem(new MenuItem("CAttackHpU", "-> If Hp <", true).SetValue(new Slider(40)));
+            }
+
+            var spells = Menu.AddSubMenu(new Menu("Dodge Spell", "Dodge Spell"));
+            {
+                foreach (var hero in
+                    HeroManager.Enemies.Where(i => Spells.Any(a => a.ChampionName == i.ChampionName)))
+                {
+                    spells.AddSubMenu(new Menu(hero.ChampionName, "ET_" + hero.ChampionName));
+                }
+
+                foreach (
+                    var spell in
+                    Spells.Where(
+                        i => HeroManager.Enemies.Any(a => a.ChampionName == i.ChampionName)))
+                {
+                    spells.SubMenu("ET_" + spell.ChampionName)
+                        .AddItem(
+                            new MenuItem(spell.MissileName, spell.ChampionName + " (" + spell.Slot + ")", true)
+                                .SetValue(true));
+                }
+            }
+
+            Menu.AddItem(new MenuItem("EvadeTargetW", "Use W", true).SetValue(true));
+            Menu.AddItem(new MenuItem("EvadeTargetE", "Use E (To Dash Behind WindWall)", true).SetValue(true));
+            Menu.AddItem(new MenuItem("EvadeTargetETower", "-> Under Tower", true).SetValue(false));
 
             Game.OnUpdate += OnUpdate;
             GameObject.OnCreate += OnCreate;
             GameObject.OnDelete += OnDelete;
             Obj_AI_Base.OnProcessSpellCast += OnProcessSpellCast;
+        }
+
+        private static void InitSpells()
+        {
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Ahri", SpellNames = new[] { "ahrifoxfiremissiletwo" }, Slot = SpellSlot.W });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Ahri", SpellNames = new[] { "ahritumblemissile" }, Slot = SpellSlot.R });
+            Spells.Add(
+                new SpellData { ChampionName = "Akali", SpellNames = new[] { "akalimota" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData { ChampionName = "Anivia", SpellNames = new[] { "frostbite" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData { ChampionName = "Annie", SpellNames = new[] { "disintegrate" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Brand",
+                    SpellNames = new[] { "brandconflagrationmissile" },
+                    Slot = SpellSlot.E
+                });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Brand",
+                    SpellNames = new[] { "brandwildfire", "brandwildfiremissile" },
+                    Slot = SpellSlot.R
+                });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Caitlyn",
+                    SpellNames = new[] { "caitlynaceintheholemissile" },
+                    Slot = SpellSlot.R
+                });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Cassiopeia", SpellNames = new[] { "cassiopeiatwinfang" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData { ChampionName = "Elise", SpellNames = new[] { "elisehumanq" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Ezreal",
+                    SpellNames = new[] { "ezrealarcaneshiftmissile" },
+                    Slot = SpellSlot.E
+                });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "FiddleSticks",
+                    SpellNames = new[] { "fiddlesticksdarkwind", "fiddlesticksdarkwindmissile" },
+                    Slot = SpellSlot.E
+                });
+            Spells.Add(
+                new SpellData { ChampionName = "Gangplank", SpellNames = new[] { "parley" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData { ChampionName = "Janna", SpellNames = new[] { "sowthewind" }, Slot = SpellSlot.W });
+            Spells.Add(
+                new SpellData { ChampionName = "Kassadin", SpellNames = new[] { "nulllance" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Katarina",
+                    SpellNames = new[] { "katarinaq", "katarinaqmis" },
+                    Slot = SpellSlot.Q
+                });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Kayle", SpellNames = new[] { "judicatorreckoning" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Leblanc",
+                    SpellNames = new[] { "leblancchaosorb", "leblancchaosorbm" },
+                    Slot = SpellSlot.Q
+                });
+            Spells.Add(new SpellData { ChampionName = "Lulu", SpellNames = new[] { "luluw" }, Slot = SpellSlot.W });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Malphite", SpellNames = new[] { "seismicshard" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "MissFortune",
+                    SpellNames = new[] { "missfortunericochetshot", "missFortunershotextra" },
+                    Slot = SpellSlot.Q
+                });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Nami",
+                    SpellNames = new[] { "namiwenemy", "namiwmissileenemy" },
+                    Slot = SpellSlot.W
+                });
+            Spells.Add(
+                new SpellData { ChampionName = "Nunu", SpellNames = new[] { "iceblast" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData { ChampionName = "Pantheon", SpellNames = new[] { "pantheonq" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Ryze",
+                    SpellNames = new[] { "spellflux", "spellfluxmissile" },
+                    Slot = SpellSlot.E
+                });
+            Spells.Add(
+                new SpellData { ChampionName = "Shaco", SpellNames = new[] { "twoshivpoison" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData { ChampionName = "Shen", SpellNames = new[] { "shenvorpalstar" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData { ChampionName = "Sona", SpellNames = new[] { "sonaqmissile" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData { ChampionName = "Swain", SpellNames = new[] { "swaintorment" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData { ChampionName = "Syndra", SpellNames = new[] { "syndrar" }, Slot = SpellSlot.R });
+            Spells.Add(
+                new SpellData { ChampionName = "Taric", SpellNames = new[] { "dazzle" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData { ChampionName = "Teemo", SpellNames = new[] { "blindingdart" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Tristana", SpellNames = new[] { "detonatingshot" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "TwistedFate", SpellNames = new[] { "bluecardattack" }, Slot = SpellSlot.W });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "TwistedFate", SpellNames = new[] { "goldcardattack" }, Slot = SpellSlot.W });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "TwistedFate", SpellNames = new[] { "redcardattack" }, Slot = SpellSlot.W });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Urgot",
+                    SpellNames = new[] { "urgotheatseekinghomemissile" },
+                    Slot = SpellSlot.Q
+                });
+            Spells.Add(
+                new SpellData { ChampionName = "Vayne", SpellNames = new[] { "vaynecondemn" }, Slot = SpellSlot.E });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Veigar", SpellNames = new[] { "veigarprimordialburst" }, Slot = SpellSlot.R });
+            Spells.Add(
+                new SpellData
+                { ChampionName = "Viktor", SpellNames = new[] { "viktorpowertransfer" }, Slot = SpellSlot.Q });
+            Spells.Add(
+                new SpellData
+                {
+                    ChampionName = "Vladimir",
+                    SpellNames = new[] { "vladimirtidesofbloodnuke" },
+                    Slot = SpellSlot.E
+                });
         }
 
         private static bool GoThroughWall(Vector2 pos1, Vector2 pos2)
@@ -84,7 +275,7 @@ namespace Flowers_Yasuo.Evade
 
             var unit = (AIHeroClient)missile.SpellCaster;
             var spellData =
-                SpellManager.Spells.FirstOrDefault(
+                Spells.FirstOrDefault(
                     i =>
                     i.SpellNames.Contains(missile.SData.Name.ToLower())
                     && Menu.Item(i.MissileName, true) != null
