+++ Core/Champion Ports/Yasuo/Flowers' Yasuo/Manager/Events/Games/Mode/Auto.cs
@@ -2,6 +2,7 @@ using EloBuddy;
 using LeagueSharp.Common; 
 namespace Flowers_Yasuo.Manager.Events.Games.Mode
 {
+    using Common;
     using System.Linq;
     using Spells;
     using LeagueSharp;
@@ -18,9 +19,11 @@ namespace Flowers_Yasuo.Manager.Events.Games.Mode
                     HeroManager.Enemies.Where(x => x.IsValidTarget(R.Range))
                         .Count(x => x.HasBuffOfType(BuffType.Knockup) || x.HasBuffOfType(BuffType.Knockback));
 
+                var Allies = HeroManager.Allies.Count(x => x.DistanceToPlayer() <= R.Range);
+
                 if (Enemies >= Menu.Item("AutoRCount", true).GetValue<Slider>().Value &&
                     Me.HealthPercent >= Menu.Item("AutoRMyHp", true).GetValue<Slider>().Value &&
-                    Me.CountEnemiesInRange(1500) <= Menu.Item("AutoRRangeCount", true).GetValue<Slider>().Value)
+                    Allies >= Menu.Item("AutoRRangeCount", true).GetValue<Slider>().Value)
                 {
                     R.Cast();
                 }
